Revert "separated profile page into separate pages"
