+++ app/(tabs)/_layout.tsx
@@ -9,59 +9,77 @@ import { useUserContext } from "@/context/UserContext";
 import { useTheme } from "react-native-paper";
 
 export default function TabLayout() {
-    const { user, accountType } = useUserContext();
-    const { colors } = useTheme();
+  const { user, accountType } = useUserContext();
+  const { colors } = useTheme();
 
-    if (!accountType) {
-        return <Redirect href="/(auth)/main" />;
-    }
+  if (!accountType) {
+    return <Redirect href="/(auth)/main" />;
+  }
 
-    return (
-        <Tabs
-            screenOptions={{
-                headerShown: false,
-                tabBarButton: HapticTab,
-                tabBarBackground: TabBarBackground,
-                tabBarStyle: {
-                    backgroundColor: colors.surface,
-                    position: Platform.OS === "ios" ? "absolute" : "relative",
-                },
-            }}
-        >
-            <Tabs.Screen
-                name="opportunity"
-                options={{
-                    tabBarLabel: () => null,
-                    tabBarIcon: ({ focused }) => (
-                        <Feather name="briefcase" size={30} color={focused ? colors.tabIconSelected : colors.tabIconDefault} />
-                    ),
-                }}
+  return (
+    <Tabs
+      screenOptions={{
+        headerShown: false,
+        tabBarButton: HapticTab,
+        tabBarBackground: TabBarBackground,
+        tabBarStyle: {
+          backgroundColor: colors.surface,
+          position: Platform.OS === "ios" ? "absolute" : "relative",
+        },
+      }}
+    >
+      <Tabs.Screen
+        name="opportunity"
+        options={{
+          tabBarLabel: () => null,
+          tabBarIcon: ({ focused }) => (
+            <Feather
+              name="briefcase"
+              size={30}
+              color={focused ? colors.tabIconSelected : colors.tabIconDefault}
             />
-            <Tabs.Screen
-                name="explore"
-                options={{
-                    tabBarLabel: () => null,
-                    tabBarIcon: ({ focused }) => (
-                        <Feather name="book-open" size={30} color={focused ? colors.tabIconSelected : colors.tabIconDefault} />
-                    ),
-                }}
+          ),
+        }}
+      />
+      <Tabs.Screen
+        name="explore"
+        options={{
+          tabBarLabel: () => null,
+          tabBarIcon: ({ focused }) => (
+            <Feather
+              name="book-open"
+              size={30}
+              color={focused ? colors.tabIconSelected : colors.tabIconDefault}
             />
-            <Tabs.Screen
-                name="chat"
-                options={{
-                    tabBarLabel: () => null,
-                    tabBarIcon: ({ focused }) => (
-                        <Feather name="message-square" size={30} color={focused ? colors.tabIconSelected : colors.tabIconDefault} />
-                    ),
-                }}
+          ),
+        }}
+      />
+      <Tabs.Screen
+        name="chat"
+        options={{
+          tabBarLabel: () => null,
+          tabBarIcon: ({ focused }) => (
+            <Feather
+              name="message-square"
+              size={30}
+              color={focused ? colors.tabIconSelected : colors.tabIconDefault}
             />
-            <Tabs.Screen
-                name="account"
-                options={{
-                    tabBarLabel: () => null,
-                    tabBarIcon: ({ focused }) => <AntDesign name="user" size={30} color={focused ? colors.tabIconSelected : colors.tabIconDefault} />,
-                }}
+          ),
+        }}
+      />
+      <Tabs.Screen
+        name="index"
+        options={{
+          tabBarLabel: () => null,
+          tabBarIcon: ({ focused }) => (
+            <AntDesign
+              name="user"
+              size={30}
+              color={focused ? colors.tabIconSelected : colors.tabIconDefault}
             />
-        </Tabs>
-    );
+          ),
+        }}
+      />
+    </Tabs>
+  );
 }
